+++ src/ResourceManager/Automation/Commands.Automation/Cmdlet/GetAzureAutomationDscCompilationJobOutput.cs
@@ -23,7 +23,7 @@ namespace Microsoft.Azure.Commands.Automation.Cmdlet
     /// <summary>
     /// Gets stream for a compilation job
     /// </summary>
-    [Cmdlet(VerbsCommon.Get, "AzureRmAutomationDscCompilationJobOutput")]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "AutomationDscCompilationJobOutput")]
     [OutputType(typeof(JobStream))]
     public class GetAzureAutomationDscCompilationJobOutput : AzureAutomationBaseCmdlet
     {