+++ src/ResourceManager/AzureBatch/Commands.Batch/Files/GetBatchNodeFileContentCommand.cs
@@ -20,7 +20,7 @@ using Constants = Microsoft.Azure.Commands.Batch.Utils.Constants;
 
 namespace Microsoft.Azure.Commands.Batch
 {
-    [Cmdlet(VerbsCommon.Get, Constants.AzureBatchNodeFileContent), OutputType(typeof(void))]
+    [Cmdlet("Get", ResourceManager.Common.AzureRMConstants.AzurePrefix + "BatchNodeFileContent"), OutputType(typeof(void))]
     public class GetBatchNodeFileContentCommand : BatchObjectModelCmdletBase
     {
         internal const string TaskAndIdAndPathParameterSet = "Task_Id_Path";