+++ src/ResourceManager/Compute/Commands.Compute/Extension/DSC/PublishAzureVMDscConfigurationCommand.cs
@@ -1,4 +1,4 @@
-using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
+﻿using Microsoft.Azure.Commands.Common.Authentication.Abstractions;
 using Microsoft.Azure.Commands.Common.Authentication.Models;
 using Microsoft.Azure.Commands.Compute.Common;
 using Microsoft.Azure.Commands.ResourceManager.Common.ArgumentCompleters;
@@ -15,13 +15,7 @@ namespace Microsoft.Azure.Commands.Compute.Extension.DSC
     /// later can be applied to Azure Virtual Machines using the 
     /// Set-AzureRmVMDscExtension cmdlet.
     /// </summary>
-    [Cmdlet(
-        VerbsData.Publish,
-        ProfileNouns.VirtualMachineDscConfiguration,
-        SupportsShouldProcess = true,
-        DefaultParameterSetName = UploadArchiveParameterSetName),
-    OutputType(
-         typeof(String))]
+    [Cmdlet("Publish", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "VMDscConfiguration",SupportsShouldProcess = true,DefaultParameterSetName = UploadArchiveParameterSetName),OutputType(typeof(String))]
     public class PublishAzureVMDscConfigurationCommand : DscExtensionPublishCmdletCommonBase
     {
         private const string CreateArchiveParameterSetName = "CreateArchive";
@@ -206,4 +200,3 @@ namespace Microsoft.Azure.Commands.Compute.Extension.DSC
         }
     }
 }
-