+++ src/ResourceManager/DataMigration/Commands.DataMigration/Cmdlets/NewProjectCmdlet.cs
@@ -25,7 +25,7 @@ namespace Microsoft.Azure.Commands.DataMigration.Cmdlets
     /// <summary>
     /// Class that creates a new instance DMS Project.
     /// </summary>
-    [Cmdlet(VerbsCommon.New, "AzureRmDataMigrationProject", DefaultParameterSetName = ComponentNameParameterSet, SupportsShouldProcess = true), OutputType(typeof(PSProject))]
+    [Cmdlet("New", ResourceManager.Common.AzureRMConstants.AzureRMPrefix + "DataMigrationProject", DefaultParameterSetName = ComponentNameParameterSet, SupportsShouldProcess = true), OutputType(typeof(PSProject))]
     [Alias("New-AzureRmDmsProject")]
     public class NewProjectCmdlet : DataMigrationCmdlet
     {